Fix: Adjust `assertTrue` call in `PipelineTest::discardRows()`
When a row is marked as discarded it is not removed from the pipeline, but its data is not transformed or loaded.  The call to `assertTrue` here should confirm that the pipeline data is unchanged.

Functionality of discarding rows is tested in `EtlTest::getArrayOfEtlData`, which tests the output of the ETL as a whole, when data has been discarded.