Don't call `Pipeline::valid()` recursively when a row is discarded (#119)
* Don't call `Pipeline::valid()` recursively when a row is discarded

Currently, when a row is discarded (e.g. by the Transformer), the Pipeline::valid function manually increments the current key, then calls itself recursively.  When discarding many rows, this not only leads to extremely deep call stacks, it also seems to cause execution time to grow linearly as extraction loop progresses.

By changing the discarded behaviour to a simple `break;` statement the valid function will simply return `true`, bypassing subsequent steps (e.g. the Loader) and progressing on to the next row without recursion.

* (chore) quick fix to make CI to pass and add PHP 8.3 and 8.4 in the CI (#122)

* Fix: Adjust `assertTrue` call in `PipelineTest::discardRows()`

When a row is marked as discarded it is not removed from the pipeline, but its data is not transformed or loaded.  The call to `assertTrue` here should confirm that the pipeline data is unchanged.

Functionality of discarding rows is tested in `EtlTest::getArrayOfEtlData`, which tests the output of the ETL as a whole, when data has been discarded.

* Feat: Add `PipelineTest::recursionWhenConsecutiveRowsDiscarded`

Adds a test to ensure that stack size does not grow out of control when consecutive rows are discarded.

---------

Co-authored-by: Eric COURTIAL <57905107+ecourtial@users.noreply.github.com>